+++ src/libraries/Microsoft.PowerFx.Interpreter/CustomFunction/CustomTexlFunction.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Numerics;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.PowerFx.Core.Functions;
+using Microsoft.PowerFx.Core.Localization;
+using Microsoft.PowerFx.Core.Types;
+using Microsoft.PowerFx.Core.Utils;
+using Microsoft.PowerFx.Types;
+using static Microsoft.PowerFx.Core.Localization.TexlStrings;
+
+namespace Microsoft.PowerFx
+{
+    /// <summary>
+    /// Internal adapter for adding custom functions. 
+    /// </summary>
+    internal class CustomTexlFunction : TexlFunction
+    {
+        public Func<IServiceProvider, FormulaValue[], CancellationToken, Task<FormulaValue>> _impl;
+
+        internal BigInteger LamdaParamMask;
+
+        public CustomTexlFunction(string name, FormulaType returnType, params FormulaType[] paramTypes)
+            : this(name, returnType._type, Array.ConvertAll(paramTypes, x => x._type))
+        {
+        }
+
+        public CustomTexlFunction(string name, DType returnType, params DType[] paramTypes)
+            : base(DPath.Root, name, name, SG("Custom func " + name), FunctionCategories.MathAndStat, returnType, 0, paramTypes.Length, paramTypes.Length, paramTypes)
+        {
+        }
+
+        public override bool IsSelfContained => true;
+
+        public static StringGetter SG(string text)
+        {
+            return (string locale) => text;
+        }
+
+        public override IEnumerable<TexlStrings.StringGetter[]> GetSignatures()
+        {
+            yield return new[] { SG("Arg 1") };
+        }
+
+        public virtual Task<FormulaValue> InvokeAsync(IServiceProvider serviceProvider, FormulaValue[] args, CancellationToken cancellationToken)
+        {
+            return _impl(serviceProvider, args, cancellationToken);
+        }
+
+        public override bool IsLazyEvalParam(int index)
+        {
+            return LamdaParamMask.TestBit(index);
+        }
+    }
+}
