+++ src/libraries/Microsoft.PowerFx.Interpreter/CustomFunction/FunctionDescr.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Numerics;
+using System.Reflection;
+using System.Text;
+using Microsoft.PowerFx.Types;
+
+namespace Microsoft.PowerFx
+{
+    internal class FunctionDescr
+    {
+        internal string Name { get; }
+
+        internal MethodInfo Method { get; }
+
+        internal FormulaType RetType { get; }
+
+        // User-facing parameter types. 
+        internal FormulaType[] ParamTypes { get; }
+
+        // If not null, then arg0 is from RuntimeConfig
+        internal Type ConfigType { get; }
+
+        internal bool IsAsync { get; }
+
+        internal BigInteger LamdaParamMask { get; }
+
+        public FunctionDescr(string name, MethodInfo method, FormulaType retType, FormulaType[] paramTypes, Type configType, BigInteger lamdaParamMask, bool isAsync = false)
+        {
+            Name = name;
+            Method = method;
+            RetType = retType;
+            ParamTypes = paramTypes;
+            ConfigType = configType;
+            IsAsync = isAsync;
+            LamdaParamMask = lamdaParamMask;
+        }
+    }
+}
